requirements.txt
torch
torchvision
transformers
opencv-python
ultralytics
Pillow